Add dynamic typecheck. The testsuite doesn't pass yet.
It isn't possible right now to initialize MemStateServer
of external states as types can't be parsed. This is
because during deployment, we need external states, but
the code isn't compiled yet to get type descriptors.
While `ScillaJIT::create` can be split to separate
compiling the code and initializaing, that isn't neat.
So the solution to write a `TypeMapDepthParser` class
that can independently get map-depth from a Scilla type,
and this can be used to set up `MemStateServer`.