Dynamic type checks (#22)
* Validate contract and transition parameters

* Fix bug in memory management of LLVM JIT object.

This fixes the bug of memory leak on exceptions, thus
allowing adding of tests to catch expected exception.

* Add dynamic typecheck. The testsuite doesn't pass yet.

It isn't possible right now to initialize MemStateServer
of external states as types can't be parsed. This is
because during deployment, we need external states, but
the code isn't compiled yet to get type descriptors.
While `ScillaJIT::create` can be split to separate
compiling the code and initializaing, that isn't neat.
So the solution to write a `TypeMapDepthParser` class
that can independently get map-depth from a Scilla type,
and this can be used to set up `MemStateServer`.

* Add an independent type parser to compute map depth

* Get remote_state_reads init test working

The first transition test doesn't work yet though because
we're doing dynamic typecheck of contract parameters even
when executing messages, which isn't needed. And in the case
of our tests, remote_state_reads.state_1.json doesn't contain
external field information for these contract parameters.

* Dynamic typechecks of contract parameters only on deploy

* Make map depth parser reject invalid types