Optimize Scilla interface for integer operations.
We don't directly involve `SafeInt` in the C interface
because that doesn't work, most likely because `SafeInt`
doesn't satisfy `std::is_trivially_copyable`.

There are checks in place to ensure that `SafeInt` and
`RawInt` are byte-compatible, and using `RawInt` in the
C interface works. Not being trivially copyable means
that `SafeInt` <-> `RawInt` byte conversions *may not* work,
but it seems to be working.